+++ AuthScreen/Auth/Models/Charts/Statistic.swift
