+++ AuthScreen/Auth/Models/Charts/Trends.swift
