+++ AuthScreen/Auth/Models/FootballModels/GoalsStatistic.swift
